requirements.txt
amqp==5.3.1
anyio==4.11.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
bidict==0.23.1
billiard==4.2.3

brotli==1.2.0
cbor2==5.7.1
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7.1
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-celery-beat==2.8.1
django-debug-toolbar==6.1.0
django-elasticsearch-dsl==7.4
django-elasticsearch-dsl-drf==0.22.5
django-filter==25.2
django-nine==0.2.7
django-redis==6.0.0
django-rest-framework-nested==0.0.1
django-timezone-field==7.1
djangorestframework==3.15.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-nested-routers==0.94.1
drf-orjson-renderer==1.7.3
drf-spectacular==0.29.0
elastic-transport==9.2.0
elasticsearch==7.17.12
elasticsearch-dsl==7.4.1
eventlet==0.40.3




h11==0.16.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
iniconfig==2.3.0


jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4

MarkupSafe==3.0.3

mysqlclient==2.2.7
orjson==3.11.4
packaging==25.0
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
prompt_toolkit==3.0.52
psutil==7.1.3
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodomex==3.23.0
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyOpenSSL==25.3.0
pytest==9.0.1
python-alipay-sdk==3.4.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-socketio==5.15.0
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
requests==2.32.4
rpds-py==0.29.0
service-identity==24.2.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.9.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.14
websocket-client==1.9.0
django-cors-headers==4.9.0
wsproto==1.3.2
zope.event==6.1
zope.interface==8.1.1
gunicorn==21.2.0
django-environ==0.12.0